Default data cache location is per-user (#157)
The default cache location on linux is now /tmp/servicex_<user>, and on windows similar, but in the TEMP directory (which is already per-user).
Also updated the docs to note that on linux the /etc/servicex directory is searched for a file since the confuse python config library is getting used.
Fixes #147