Delete Dataset From CLI (#519)
* Add command line to get a particular dataset and list the physical files

* Add command to delete a specific dataset from the cache

* Allow users to view deleted datasets

* Migrate dataset methods in ServiceX Adapter to use the aiohttp library
This follows the examples used by the transform methods and allows us to
use the _extract_message function to get better error messages.

During development of this I learned that in unit tests that test for
exceptions, any assertions inside the pytest.raises context manager
will not be executed! Fixed numerous examples of that.